requirements.txt
twitchio==2.8.2
deep-translator==1.11.4
python-dotenv==1.0.0
psutil==5.9.8
aiohttp==3.9.3
transformers==4.36.2
torch==2.1.2
sentencepiece==0.1.99
protobuf==4.25.1
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1
langdetect==1.0.9
emoji==2.8.0
